// src/modules/inventory/reports/yarn-store/partywise-yarnissue-greyrcv-balance-summary-report/partyw-yarni-greyr-balance-sum-rpt-type.tsx
export type PartyWiseYanrIssueAndGreyRcvSummaryType = {
  COMPANY_ID: number;
  COMPANY_NAME: string;
  COMPANY_ADDRESS: string;
  KNITTING_HOUSE_ID: number;
  KNIT_HOUSE: string;
  YARN_ISSUE_QTY: number;
  GREY_RECEIVE_QTY: number;
  KNT_PROCESS_LOSS_QTY: number;
  YARN_BALANCE: number;
  YARN_BALANCE_VALUE: number;
};